# Increase Performance of \ (and /) for small square matrix