--- a/src/solve.jl
+++ b/src/solve.jl
@@ -26,6 +26,22 @@ end
             (a[1,1]*a[2,2] - a[1,2]*a[2,1])*b[3]) / d )
 end
 
+for Dim in 2:3  # no improvements for Dim = 1 and if no specialized solve is available
+    @eval begin
+    @inline function solve(::Size{($Dim,$Dim)}, ::Size{Sb}, a::StaticMatrix{<:Any, <:Any, Ta}, b::StaticMatrix{<:Any, <:Any, Tb}) where {Sb, Ta, Tb}
+        d = det(a)
+        T = typeof((one(Ta)*zero(Tb) + one(Ta)*zero(Tb))/d)
+        c = similar(b, T)
+        for col = 1:Sb[2]
+            @inbounds c[:, col] = solve(Size($Dim,$Dim), Size($Dim,), a, b[:, col])
+        end
+        return c
+    end
+    end # @eval
+end
+
+
+
 @generated function solve(::Size{Sa}, ::Size{Sb}, a::StaticMatrix{<:Any, <:Any, Ta}, b::StaticVecOrMat{Tb}) where {Sa, Sb, Ta, Tb}
     if Sa[end] != Sb[1]
         throw(DimensionMismatch("right hand side B needs first dimension of size $(Sa[end]), has size $Sb"))