Concerning [this issue](https://github.com/JuliaArrays/StaticArrays.jl/issues/582#issue-398546390).
My proposal for increased performance of A/B if A is 2x2 or 3x3 and B is 2or3xn.
There is an [ongoing discussion](https://discourse.julialang.org/t/performance-of-matrix-division/19477?u=judober) about this.