@judober 2019-01-14
Speed comparison (other machine than on discourse).
Original version:
```
julia> a = Float64.(@SMatrix rand(3,3)); b = Float64.(@SMatrix rand(3,3));

julia> @benchmark a\b
BenchmarkTools.Trial: 
  memory estimate:  80 bytes
  allocs estimate:  1
  --------------
  minimum time:     149.989 ns (0.00% GC)
  median time:      153.293 ns (0.00% GC)
  mean time:        184.259 ns (7.69% GC)
  maximum time:     75.550 μs (99.69% GC)
  --------------
  samples:          10000
  evals/sample:     818
```

This PR (with the same matrices):
```
julia> @benchmark a\b
BenchmarkTools.Trial: 
  memory estimate:  80 bytes
  allocs estimate:  1
  --------------
  minimum time:     66.859 ns (0.00% GC)
  median time:      70.639 ns (0.00% GC)
  mean time:        98.225 ns (16.09% GC)
  maximum time:     73.637 μs (99.82% GC)
  --------------
  samples:          10000
  evals/sample:     978
```