@tkoolen 2019-01-14
Could you run the following benchmark (results below are on my machine with StaticArrays 0.10.2):

```julia
julia> using StaticArrays, BenchmarkTools

julia> @benchmark a \ b setup = begin
           a = rand(SMatrix{3, 3})
           b = rand(SMatrix{3, 3})
       end
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     71.394 ns (0.00% GC)
  median time:      71.689 ns (0.00% GC)
  mean time:        74.010 ns (0.00% GC)
  maximum time:     113.456 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     974
```