@andyferris 2019-01-15
Seems like a reasonable approach to me.