@judober 2019-01-15
New test with old version:
```
julia> using StaticArrays, BenchmarkTools

julia> @benchmark a \ b setup = begin
                  a = rand(SMatrix{3, 3})
                  b = rand(SMatrix{3, 3})
              end
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     104.663 ns (0.00% GC)
  median time:      106.303 ns (0.00% GC)
  mean time:        118.250 ns (0.00% GC)
  maximum time:     1.198 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     937
```

With PR:
```
julia> using StaticArrays, BenchmarkTools

julia> @benchmark a \ b setup = begin
                  a = rand(SMatrix{3, 3})
                  b = rand(SMatrix{3, 3})
              end
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     30.143 ns (0.00% GC)
  median time:      32.320 ns (0.00% GC)
  mean time:        35.998 ns (0.00% GC)
  maximum time:     1.324 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     994
```