@tkoolen 2019-01-15
Do you think the repeated computation of the determinant of `A` for each of the columns inside the inner `solve` calls is being optimized away? How long does `det(A)` take on your machine?