@judober 2019-01-15
Ok, first the performance of det(A):
```
julia> using LinearAlgebra

julia> @benchmark det(a) setup = begin
                  a = rand(SMatrix{3, 3})
              end
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     17.886 ns (0.00% GC)
  median time:      17.888 ns (0.00% GC)
  mean time:        20.050 ns (0.00% GC)
  maximum time:     169.474 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     998
```
You might be right that this is a point for further improvements. I will try it.