@judober 2019-01-15
The results when adding OneTo:
```
julia> @benchmark a \ b setup = begin
                  a = rand(SMatrix{3, 3})
                  b = rand(SMatrix{3, 3})
              end
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     30.143 ns (0.00% GC)
  median time:      30.303 ns (0.00% GC)
  mean time:        34.886 ns (0.00% GC)
  maximum time:     460.771 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     994
```
seems pretty identical.