@tkoolen 2019-01-16
> Though, when I compare the numbers, it must be optimised away, otherwise the four calculations of det(a) would need alone approx. 60 ns.

Yeah, I agree. That's cool!

No further comments from me, @andyferris, good to go?